Need help with calendar bug
Refactored the code to use Observable architecture
Cleaned up the calendar formatting with dedicated functions
