+++ lib/Screens/Calendar/Calendar.dart
@@ -1,0 +1,262 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:intl/intl.dart';
+import 'package:table_calendar/table_calendar.dart';
+import 'package:untitled3/Model/CalendarEvent.dart';
+import 'package:untitled3/Model/Note.dart';
+import 'package:untitled3/Observables/CalenderObservable.dart';
+import 'package:untitled3/Observables/SettingObservable.dart';
+import 'package:untitled3/Observables/NoteObservable.dart';
+import 'package:provider/provider.dart';
+import 'package:untitled3/Screens/Calendar/CalendarFormatBar.dart';
+
+final viewCalendarScaffoldKey = GlobalKey<ScaffoldState>();
+
+//Variable Definitions ----------------------------------------
+bool _filteredNotesIsVisible = false;
+
+DateTime _focusedDay = DateTime.now();
+
+ValueNotifier<bool> _dayAndEventsUpdated = ValueNotifier(false);
+
+//--------------------------------------------------------------------------------
+
+class Calendar extends StatefulWidget {
+  @override
+  CalendarState createState() => CalendarState();
+}
+
+class CalendarState extends State<Calendar> {
+  get calendarObserver => this.calendarObserver;
+  List<TextNote> _matchedEvents = [];
+
+  @override
+  Widget build(BuildContext context) {
+    final noteObserver = Provider.of<NoteObserver>(context);
+    final calendarObserver = Provider.of<CalendarObservable>(context);
+    final settingObserver = Provider.of<SettingObserver>(context);
+
+    calendarObserver.setNoteObserver(noteObserver);
+    calendarObserver.generateDailyTiles();
+    List<Container> _daysAndEvents = calendarObserver.generateDailyTiles();
+
+    // This function is called whenever the text field changes
+    void _runFilter(String value) {
+      final List<TextNote> _allEvents = noteObserver.usersNotes;
+      List<TextNote> _results = [];
+      if ((value.isEmpty)) {
+        setState(() {
+          calendarObserver.changeFormat(CalendarFormat.month);
+          calendarObserver.setCalendarBarIsVisible(true);
+
+          calendarObserver.setMonthHasBeenPressed(true);
+          calendarObserver.setWeekHasBeenPressed(false);
+          calendarObserver.setDayHasBeenPressed(false);
+
+          calendarObserver.setCalendarIsVisible(true);
+
+          _filteredNotesIsVisible = false;
+        });
+      } else {
+        _results = _allEvents
+            .where((event) =>
+                event.text.toLowerCase().contains(value.toLowerCase()))
+            .toList();
+        // Refresh the UI
+        setState(() {
+          _matchedEvents = _results;
+          _filteredNotesIsVisible = true;
+          calendarObserver.setCalendarBarIsVisible(false);
+          calendarObserver.setCalendarIsVisible(false);
+          calendarObserver.setDailyCalendarIsVisible(false);
+        });
+      }
+    }
+
+    return Observer(
+      builder: (_) => Padding(
+        padding: const EdgeInsets.all(10.0),
+        child: Column(children: <Widget>[
+          TextField(
+              onChanged: (value) => _runFilter(value),
+              decoration: InputDecoration(
+                border: OutlineInputBorder(),
+                hintText: '--Search For A Note--',
+              )),
+          const SizedBox(
+            height: 15,
+          ),
+          Visibility(
+            visible: _filteredNotesIsVisible,
+            child: Expanded(
+              child: _matchedEvents.isNotEmpty
+                  ? new ListView.builder(
+                      itemCount: _matchedEvents.length,
+                      itemBuilder: (context, index) => new Container(
+                        key: ValueKey(_matchedEvents[index].noteId),
+                        margin: const EdgeInsets.symmetric(
+                          horizontal: 12.0,
+                          vertical: 4.0,
+                        ),
+                        decoration: BoxDecoration(
+                          color: Colors.lightBlue.shade50,
+                          border: Border.all(color: Colors.blueGrey, width: 1),
+                          borderRadius: BorderRadius.circular(12.0),
+                        ),
+                        child: new ListTile(
+                          //onTap: () => print('${value[index]}'),
+                          title: Text("${_matchedEvents[index].text}",
+                              textAlign: TextAlign.center),
+                          subtitle: Text(
+                              " on ${DateFormat('MM-dd-yyyy').format(DateTime.parse((_matchedEvents[index].eventDate)))} \t at \t ${_matchedEvents[index].eventTime}",
+                              textAlign: TextAlign.center),
+                        ),
+                      ),
+                    )
+                  : const Text(
+                      'No results found',
+                      style: TextStyle(fontSize: 24),
+                    ),
+            ),
+          ),
+
+          //Calendar Buttons "Week | Month | Day"-------------------------------
+          Visibility(
+            visible: calendarObserver.getCalendarBarIsVisible(),
+            child: calendarFormatBar(),
+          ),
+
+          //Calendar Table------------------------------------------------------
+          Visibility(
+            visible: calendarObserver.getCalendarIsVisible(),
+            child: TableCalendar(
+              //this sets the calendar Format to what is found in the observer
+              calendarFormat: calendarObserver.calendarFormat,
+              //This reaches into the TableCalendar and sets some header properties
+              headerStyle: HeaderStyle(
+                formatButtonVisible: false,
+                titleCentered: true,
+                headerPadding: EdgeInsets.all(5),
+              ),
+              focusedDay: _focusedDay,
+              locale: settingObserver.userSettings.locale.languageCode,
+              firstDay: calendarObserver
+                  .getStartDay(), //Date of the oldest past event
+              lastDay: calendarObserver.getEndDay(), //Date of the last event
+              selectedDayPredicate: (day) {
+                return isSameDay(calendarObserver.selectedDay, day);
+              },
+              //-------------------------------------------------------------------------------------------
+/*
+                //This event loader is causing an error and slowing this whole thing down
+                eventLoader: (DateTime day) {
+                  return calendarObserver
+                      .loadEventsOfSelectedDay(day.toString().split(" ")[0]);
+                },
+
+ */
+              //-----------------------------------------------------------------------------------------
+              onDaySelected: (selectedDay, focusDay) {
+                _focusedDay = selectedDay;
+                print("onDaySelected selectedDay: $selectedDay");
+                //extract the date portion
+                calendarObserver.setSelectedDay(selectedDay);
+                var _events = calendarObserver.loadEventsOfSelectedDay(
+                    selectedDay.toString().split(" ")[0]);
+
+                if (_events.length > 0) {
+                  calendarObserver.weekView();
+                  calendarObserver.getNotesOnDay();
+                }
+                (context as Element).reassemble();
+              },
+
+              onPageChanged: (focusedDay) {},
+              calendarStyle: CalendarStyle(
+                  selectedDecoration: BoxDecoration(
+                    color: Colors.pink,
+                    shape: BoxShape.circle,
+                  ),
+                  todayDecoration: BoxDecoration(
+                    color: Colors.blueAccent,
+                    shape: BoxShape.circle,
+                  ),
+                  selectedTextStyle: TextStyle(
+                      fontWeight: FontWeight.bold,
+                      fontSize: 18.0,
+                      color: Colors.white)),
+            ),
+          ),
+
+          const SizedBox(height: 8.0),
+
+          //Scrolling Daily View of the Calendar--------------------------------
+          Visibility(
+            visible: calendarObserver.getDailyCalendarIsVisible(),
+            child: Expanded(
+              child: ValueListenableBuilder<bool>(
+                  valueListenable: _dayAndEventsUpdated,
+                  builder: (context, value, _) {
+                    return ListView.builder(
+                      itemCount: _daysAndEvents.length,
+                      itemExtent: 50,
+                      controller: ScrollController(
+                          initialScrollOffset:
+                              50 * _daysAndEvents.length / 2 - 50),
+                      itemBuilder: (context, index) => new Container(
+                          margin: const EdgeInsets.symmetric(
+                            horizontal: 12.0,
+                            vertical: 4.0,
+                          ),
+                          key: ValueKey(_daysAndEvents[index]),
+                          child: _daysAndEvents[index]),
+                    );
+                  }),
+            ),
+          ),
+
+          const SizedBox(height: 8.0),
+
+          //Area under Calendar displaying notes--------------------------------
+          Visibility(
+            visible: calendarObserver.getNotesOnDayIsVisible(),
+            child: Expanded(
+              child:ValueListenableBuilder<List<CalenderEvent>>(
+                valueListenable: calendarObserver.selectedEvents,
+                builder: (context, value, _) {
+                  print("Initialized Value Notifier: ");
+                  return ListView.builder(
+                      itemCount: value.length,
+                      itemBuilder: (context, index) {
+                        return Container(
+                          height: 50,
+                          margin: const EdgeInsets.symmetric(
+                            horizontal: 12.0,
+                            vertical: 3,
+                          ),
+                          decoration: BoxDecoration(
+                            color: Colors.lightBlue.shade50,
+                            border: Border.all(
+                                color: Colors.blueGrey,
+                                width: 1
+                            ),
+                            borderRadius: BorderRadius.circular(12.0),
+
+                          ),
+                          child: ListTile(
+                            //onTap: () => print('${value[index]}'),
+                            title: Text("${value[index]} \t at \t ${value[index].time}",
+                                textAlign: TextAlign.center),
+                          ),
+                        );
+                      });
+                },
+              )
+            ),
+          )
+        ]),
+      ),
+    );
+  }
+}
